+++ src/commons/mixins/__init__.py
@@ -1,0 +1,5 @@
+from src.commons.mixins.user_group_mixin import UserGroupMixin
+
+__all__ = [
+    "UserGroupMixin"
+]
