+++ src/miet_angebot/views/listings.py
@@ -15,23 +15,14 @@ from src.miet_angebot.serializers import (
     HostRetrieveListingSerializer,
     ListingSerializer,
 )
+from src.commons.mixins import UserGroupMixin
 
 
-class ListingViewSet(ModelViewSet):
+class ListingViewSet(UserGroupMixin, ModelViewSet):
     filterset_class = ListingFilter
     search_fields = ["title", "description"]
     ordering_fields = ["price_per_day", "created_at"]
     http_method_names = ["get", "post", "put", "patch", "delete"]
-    user_group = None
-
-    def initial(self, request, *args, **kwargs):
-        if request.user.groups.filter(name="host").exists():
-            self.user_group = "host"
-        elif request.user.groups.filter(name="guest").exists():
-            self.user_group = "guest"
-        else:
-            self.user_group = None
-        super().initial(request, *args, **kwargs)
 
     def get_queryset(self):
         queryset = Listing.objects.filter()
@@ -79,6 +70,3 @@ class ListingViewSet(ModelViewSet):
             if self.action in ["list", "retrieve"]:
                 permissions = [IsAuthenticated(), CustomModelPermissions()]
         return permissions
-
-
-
